+++ packages/sdks/nextjs-sdk/src/client/index.ts
@@ -1,6 +1,6 @@
 'use client';
 
-// export most of the things from the SDK
+// Export most of the things from the SDK
 // we don't need to export AuthProvider, as it is exported from the root index.ts
 export {
 	useDescope,