+++ packages/widgets/user-profile-widget/src/lib/widget/index.ts
@@ -1,5 +1,4 @@
 import { compose } from '@descope/sdk-helpers';
-import WebComponent from '@descope/web-component';
 import { initMixin } from './mixins/initMixin/initMixin';
 
 declare const BUILD_VERSION: string;
@@ -8,6 +7,13 @@ const rootMixin = (superclass: CustomElementConstructor) =>
   class RootMixinClass extends initMixin(superclass) {
     async init() {
       await super.init?.();
+
+      const WebComponent: any =
+        customElements.get('descope-wc') ||
+        (await import('@descope/web-component').then(
+          (module) => module.default,
+        ));
+
       WebComponent.sdkConfigOverrides = {
         baseHeaders: {
           'x-descope-widget-type': 'user-profile-widget',