+++ packages/widgets/access-key-management-widget/CHANGELOG.md
@@ -2,6 +2,50 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.6](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.5...access-key-management-widget-0.3.6) (2025-01-02)
+
+
+### Bug Fixes
+
+* Widgets tenant ID attribute ([#879](https://github.com/descope/descope-js/issues/879)) ([6b67f5e](https://github.com/descope/descope-js/commit/6b67f5e92f596e8145a24ccb90db9fcecb5190ea))
+
+## [0.3.5](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.4...access-key-management-widget-0.3.5) (2024-12-24)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.6.0`
+## [0.3.4](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.3...access-key-management-widget-0.3.4) (2024-12-22)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.1`
+* `sdk-mixins` updated to version `0.5.2`
+## [0.3.3](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.2...access-key-management-widget-0.3.3) (2024-12-18)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.0`
+## [0.3.2](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.1...access-key-management-widget-0.3.2) (2024-12-18)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.5.1`
+## [0.3.1](https://github.com/descope/descope-js/compare/access-key-management-widget-0.3.0...access-key-management-widget-0.3.1) (2024-12-08)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.22.0`
+## [0.3.0](https://github.com/descope/descope-js/compare/access-key-management-widget-0.2.3...access-key-management-widget-0.3.0) (2024-12-04)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.21.0`
+* `sdk-mixins` updated to version `0.5.0`
+
+### Features
+
+* add permitted IPs for access key widget ([#852](https://github.com/descope/descope-js/issues/852)) ([b2b5bbe](https://github.com/descope/descope-js/commit/b2b5bbe943e3911236f67525f6acf347651d0d10))
+
 ## [0.2.3](https://github.com/descope/descope-js/compare/access-key-management-widget-0.2.2...access-key-management-widget-0.2.3) (2024-11-16)
 
 ### Dependency Updates