+++ packages/widgets/applications-portal-widget/CHANGELOG.md
@@ -2,6 +2,38 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.9](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.8...applications-portal-widget-0.2.9) (2024-12-24)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.6.0`
+## [0.2.8](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.7...applications-portal-widget-0.2.8) (2024-12-22)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.1`
+* `sdk-mixins` updated to version `0.5.2`
+## [0.2.7](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.6...applications-portal-widget-0.2.7) (2024-12-18)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.23.0`
+## [0.2.6](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.5...applications-portal-widget-0.2.6) (2024-12-18)
+
+### Dependency Updates
+
+* `sdk-mixins` updated to version `0.5.1`
+## [0.2.5](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.4...applications-portal-widget-0.2.5) (2024-12-08)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.22.0`
+## [0.2.4](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.3...applications-portal-widget-0.2.4) (2024-12-04)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.21.0`
+* `sdk-mixins` updated to version `0.5.0`
 ## [0.2.3](https://github.com/descope/descope-js/compare/applications-portal-widget-0.2.2...applications-portal-widget-0.2.3) (2024-11-16)
 
 ### Dependency Updates